refactor(studio): merge the worker settings tab into overview
Move the Container, Resources, and Develop locally sections onto the
overview tab and remove the Settings tab. The invocation snippets that
lived in both tabs are now a single How to call section.
